+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## Version 1.1.0
+## Version 1.2.0
+- Added metric aggregation for a project porfolio. Aggregation algorithms implement (https://docs.sonarqube.org/latest/user-guide/portfolios/) sonarqube aggregation logic.
 - Added capability to request metrics for multiple sonarqube projects.
 - Added capabilities to test if a sonarqube user exists and to update user name & email.